Extract non-package based plugin initialisation
